indo

+++ inc/funcionario.js
@@ -211,11 +211,11 @@ module.exports = {
                     } else {
                         resolve(results);
 
-                        if (results[0].photo != `arquivos/imgpadrao.jpg`  & results[0].photo != null ) {
+                        if (results[0].photo != `arquivos/imgpadrao.jpg`  && results[0].photo != null ) {
 
                             caminho = `./public/${results[0].photo}`
 
-                            if (caminho) {
+                            if (caminho != null) {
                                 fs.unlinkSync(caminho);
                             }
